src/training/pretraining/requirements.txt
absl-py==2.0.0
accelerate==0.24.1
attrs==23.1.0
boto3==1.29.1
botocore==1.32.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==2.2.1
contextlib2==21.6.0
dill==0.3.7
discrete-key-value-bottleneck-pytorch==0.1.1
einops==0.7.0
enformer-pytorch==0.8.7
filelock==3.13.1
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lightning-utilities==0.9.0
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
pathos==0.3.1
platformdirs==4.0.0
polars==0.19.12
pox==0.3.3
ppft==1.7.6.7
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfaidx==0.7.2.2
pysam==0.22.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.0
sagemaker==2.197.0
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sympy==1.12
tblib==1.7.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tokenizers==0.14.1
torch==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vector-quantize-pytorch==1.11.5
Werkzeug==3.0.1
zipp==3.17.0
umap-learn==0.5.3